+++ editprofile.php
@@ -3,13 +3,12 @@ require 'settings/conn.php';
 
 session_start();
 
-// Check if file is provided
 if (isset($_FILES['fileInput']) && $_FILES['fileInput']['error'] == UPLOAD_ERR_OK) {
     $tmpFilePath = $_FILES['fileInput']['tmp_name'];
     $fileName = $_FILES['fileInput']['name'];
     $ftpUploadDirectory = "/avatars/";
 
-    $uniqueFileName = uniqid() . '_' . $_SESSION['userid'] . '.' . pathinfo($fileName, PATHINFO_EXTENSION); // Use the file's original extension
+    $uniqueFileName = uniqid() . '_' . $_SESSION['userid'] . '.' . pathinfo($fileName, PATHINFO_EXTENSION);
     $destinationPath = $ftpUploadDirectory . $uniqueFileName;
 
     if (ftp_put($ftp_conn, $destinationPath, $tmpFilePath, FTP_BINARY)) {
@@ -27,11 +26,8 @@ if (isset($_FILES['fileInput']) && $_FILES['fileInput']['error'] == UPLOAD_ERR_O
     }
 }
 
-// Check if signature is provided
 if (isset($_POST['signature'])) {
     $signature = $_POST['signature'];
-
-    // Update signature in the database
     $updateSignatureQuery = "UPDATE users SET signature = ?, updated = CURRENT_TIMESTAMP() WHERE userid = ?";
     $stmt = mysqli_prepare($conn, $updateSignatureQuery);
     $userId = $_SESSION['userid'];