+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [0.1.1] - 2024-04-07
+## [0.1.2] - 2024-04-07
  
 ### Version Notes
 
 - Established FTP Connection
-- Easy way to show FTP images
+- Register and login system
+- show Pokemon Party Data
+- Dummy add pokemon and level